Refactor Filter  / CircleList now own component
